// client/mobile/src/Lib/theme.ts
export enum COLOR {
  GREEN = '#3c7548',
  SOFT_GREEN = '#3C754861',
  SOFT_GREY = '#dbdbdb',
  BG_INPUT = '#DBDBDB45',
  GREY = '#5C6360',
  BOLD_GREY = '#00000061',
  BLACK = '#6a706e',
  SOFT_BLACK = '#5C6360',
  PINK = '#DCA8B8',
  WHITE = '#FFFFFF',
  RED = '#f70000',
  SOFT_RED = '#FF5A5F',
}

export enum FONTS {
  REGULAR = 'Poppins-Regular',
  MEDIUM = 'Poppins-Medium',
  SEMI_BOLD = 'Poppins-SemiBold',
  BOLD = 'Poppins-Bold',
}

export enum LETTER_SPACING {
  LETTER_SPACING = -0.19,
}